Updated the whole documentation of the library
Updated to Doxygen 1.8.16.

Functions ending with '_rational' have the more complete documentation,
and their corresponding 'double' counterparts refer to them.

Removed the namespace "conversion" and moved it to "utils" since it is
not distributed.

Some .cpp files had 'std::vector', 'std::string', this is no longer needed.

+++ lal/linarr/C_dyn_prog.cpp
@@ -144,7 +144,7 @@ inline uint64_t __compute_crossings_dyn_prog(
 
 	/* compute number of crossings */
 
-	uint32_t C = 0;
+	uint64_t C = 0;
 
 	for (uint32_t pu = 0; pu < n - 3; ++pu) {
 		const node u = T[pu];
